updated how rate limiting is implemented
previously there was a pool of 300/600 requests that would recover slowly (one per second) so users could do bursts of requests but not use the api for scraping.

Now, there is a daily and an hourly pool. Automatically requests come out of the hourly pool until it's depleted and only then out of the daily limit be used.
When both limits are exceeded, the user can't do any requests until the next full hour.

In the past, api requests would be delayed once the limit is exceeded until the pool had recovered sufficiently, but in extreme cases this could make the client extremely slow and the user wouldn't know why.
Now, requests are actually rejected with a RateLimitError - which may be reported by the client.